Add a buncha cards (#35)
* Matt Morgan -> Orange President

* Adds cards

* Add rust + hicock

* Collect cards dynamically

* fix tests

* formatting
