+++ kaori/plugins/gacha/engine/core/__init__.py
@@ -3,5 +3,3 @@ from __future__ import annotations
 from .card import Card
 from .combat_strategies import *
 from .core import *
-
-
