+++ kaori/plugins/gacha/engine/test/__init__.py
@@ -1,2 +1,2 @@
-from .card_simulator import run_card_simulator
 from .battle_simulator import run_battle_simulator
+from .card_simulator import run_card_simulator
