+++ kaori/plugins/gacha/engine/test/dmg_cards.py
@@ -1,5 +1,4 @@
-from kaori.plugins.gacha.engine import *
-from kaori.plugins.gacha.engine import Card
+from ..core import *
 
 dmg_no_invest = Card(
     name="Baseline DMG with no investment",
