+++ kaori/plugins/gacha/engine/test/interesting_cards.py
@@ -1,2 +1,1 @@
-from .meme_cards import *
-from .balanced_cards import *
+
