+++ kaori/plugins/gacha/engine/test/meme_cards.py
@@ -1,32 +1,85 @@
-from kaori.plugins.gacha.engine import *
-from kaori.plugins.gacha.engine import Card
+from ..core import *
 
-matt_morg = Card(name="Matt Morgan",
-                 rarity=S,
-                 nature=(cursed, clown),
-                 stupid=1,
-                 baby=6,
-                 clown=3,
-                 horny=1,
-                 cursed=7,
-                 feral=2)
+orange_president = Card(
+    name="Orange President",
+    rarity=S,
+    nature=(cursed, clown),
+    stupid=1,
+    baby=6,
+    clown=3,
+    horny=1,
+    cursed=7,
+    feral=2
+)
 
-kim_jong_un_lil_sis = Card(name="Kim Jong Un's Little Sister",
-                           rarity=S,
-                           nature=(horny, baby),
-                           stupid=3,
-                           baby=6,
-                           clown=1,
-                           horny=7,
-                           cursed=2,
-                           feral=1)
+leg_hair = Card(
+    name="Leg hAir",
+    rarity=S,
+    nature=(cursed, clown),
+    stupid=3,
+    baby=6,
+    clown=1,
+    horny=6,
+    cursed=2,
+    feral=2
+)
 
-tlp = Card(name="TLP",
-           rarity=C,
-           nature=(horny, feral),
-           stupid=1,
-           baby=2,
-           clown=3,
-           horny=5,
-           cursed=2,
-           feral=3)
+kim_jong_un_lil_sis = Card(
+    name="Kim Jong Un's Little Sister",
+    rarity=S,
+    nature=(horny, baby),
+    stupid=3,
+    baby=6,
+    clown=1,
+    horny=7,
+    cursed=2,
+    feral=1
+)
+
+matt_morgan = Card(
+    name="Matt Morgan",
+    rarity=S,
+    nature=(cursed, baby),
+    stupid=2,
+    baby=6,
+    clown=1,
+    horny=2,
+    cursed=7,
+    feral=2
+)
+
+tlp = Card(
+    name="TLP",
+    rarity=C,
+    nature=(horny, feral),
+    stupid=1,
+    baby=2,
+    clown=3,
+    horny=5,
+    cursed=2,
+    feral=3
+)
+
+hicock45 = Card(
+    name="hicock45",
+    rarity=B,
+    nature=(horny, feral),
+    stupid=2,
+    baby=3,
+    clown=3,
+    horny=4,
+    cursed=1,
+    feral=6
+)
+
+the_rust_programming_language = Card(
+    name="The Rust Programming Language",
+    rarity=B,
+    nature=(horny, feral),
+    stupid=3,
+    baby=2,
+    clown=1,
+    horny=5,
+    cursed=4,
+    feral=4,
+)
