+++ kaori/plugins/gacha/engine/utils.py
@@ -1,4 +1,3 @@
-import sys
 from functools import partial
 from typing import Tuple, Union
